requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
mysql-connector-python==8.1.0
lxml==4.9.3
aiohttp==3.8.5